+++ Sources/libcoverage/include/libcoverage.h
@@ -23,13 +23,13 @@
 
 // Tracks a set of edges by their indices
 struct edge_set {
-    uint64_t count;
+    uint32_t count;
     uint32_t * edge_indices;
 };
 
 // Tracks the hit count of all edges
 struct edge_counts {
-    uint64_t count;
+    uint32_t count;
     uint32_t * edge_hit_count;
 };
 
@@ -55,13 +55,13 @@ struct cov_context {
     uint8_t* crash_bits;
 
     // Total number of edges in the target program.
-    uint64_t num_edges;
+    uint32_t num_edges;
     
     // Number of used bytes in the shmem->edges bitmap, roughly num_edges / 8.
-    uint64_t bitmap_size;
+    uint32_t bitmap_size;
     
     // Total number of edges that have been discovered so far.
-    uint64_t found_edges;
+    uint32_t found_edges;
 
 #if defined(_WIN32)
     // Mapping Handle
@@ -82,12 +82,12 @@ void cov_shutdown(struct cov_context*);
 int cov_evaluate(struct cov_context* context, struct edge_set* new_edges);
 int cov_evaluate_crash(struct cov_context*);
 
-int cov_compare_equal(struct cov_context*, uint32_t* edges, uint64_t num_edges);
+int cov_compare_equal(struct cov_context*, uint32_t* edges, uint32_t num_edges);
 
 void cov_clear_bitmap(struct cov_context*);
 
 int cov_get_edge_counts(struct cov_context* context, struct edge_counts* edges);
-void cov_clear_edge_data(struct cov_context* context, uint64_t index);
+void cov_clear_edge_data(struct cov_context* context, uint32_t index);
 void cov_reset_state(struct cov_context* context);
 
 #endif
